requirements.txt
llama-index-core>=0.10.1
llama-index-embeddings-openai>=0.1.6
llama-index-llms-openai>=0.1.6
llama-index-llms-groq>=0.1.3
llama-index-vector-stores-faiss>=0.1.1
llama-index-readers-file>=0.1.6
python-dotenv>=1.0.0
flask>=3.0.0
openai>=1.12.0
groq>=0.4.0
pypdf>=4.0.1
faiss-cpu>=1.7.4
langchain-core
langchain-ollama
langchain-huggingface
llama-index-embeddings-langchain
llama-index-llms-gemini
llama-index-llms-ollama
